+++ xaeroflux-core/src/logs/wal.rs
@@ -3,8 +3,8 @@ use std::cmp::min;
 use crate::sys::*;
 pub struct Wal {
     pub mmap: memmap2::MmapMut,
-    pub offset: usize,
-    pub page_size: usize,
+    pub offset: u32,
+    pub page_size: u32,
 }
 
 impl Drop for Wal {
@@ -40,11 +40,11 @@ impl WalOps for Wal {
             let remaining = raw_buffer.len() - offset;
             // advance only min of page size and remaining
             // this is to avoid writing more than page size
-            let write_size = min(remaining, self.page_size);
-            self.mmap[self.offset..self.offset + write_size]
+            let write_size = min(remaining, self.page_size as usize);
+            self.mmap[self.offset as usize..self.offset as usize + write_size as usize]
                 .copy_from_slice(&raw_buffer[offset..offset + write_size]);
             offset += write_size;
-            self.offset += write_size;
+            self.offset += write_size as u32;
             if offset + write_size >= self.mmap.len() {
                 offset = 0;
                 self.mmap.flush().expect("Failed to flush mmap");
@@ -53,8 +53,8 @@ impl WalOps for Wal {
     }
 
     fn read(&self) -> Vec<u8> {
-        let mut buffer = Vec::with_capacity(self.offset);
-        buffer.extend_from_slice(&self.mmap[..self.offset]);
+        let mut buffer = Vec::with_capacity(self.offset as usize);
+        buffer.extend_from_slice(&self.mmap[..self.offset as usize]);
         buffer
     }
